+++ src/bl64-hlm.bash
@@ -1,0 +1,138 @@
+#######################################
+# BashLib64 / Module / Functions / Interact with HLM
+#
+# Version: 1.0.0
+#######################################
+
+#######################################
+# Add a helm repository to the local host
+#
+# Arguments:
+#   $1: repository name
+#   $2: repository source
+# Outputs:
+#   STDOUT: command output
+#   STDERR: command stderr
+# Returns:
+#   command exit status
+#######################################
+function bl64_hlm_repo_add() {
+  bl64_dbg_lib_show_function "$@"
+  local repository="${1:-${BL64_LIB_DEFAULT}}"
+  local source="${2:-${BL64_LIB_DEFAULT}}"
+
+  bl64_check_parameter 'repository' &&
+    bl64_check_parameter 'source' ||
+    return $?
+
+  bl64_dbg_app_show_info "add helm repository (${repository})"
+  bl64_hlm_run_helm \
+    repo add \
+    "$repository" \
+    "$source" ||
+    return $?
+
+  bl64_dbg_app_show_info "try to update repository from source (${source})"
+  bl64_hlm_run_helm repo update "$repository"
+
+  return 0
+}
+
+#######################################
+# Upgrade or install and existing chart
+#
+# Arguments:
+#   $1: full path to the kube/config file with cluster credentials
+#   $2: target namespace
+#   $3: chart name
+#   $4: chart source
+#   $@: arguments are passed as-is to the command
+# Outputs:
+#   STDOUT: command output
+#   STDERR: command stderr
+# Returns:
+#   command exit status
+#######################################
+function bl64_hlm_chart_upgrade() {
+  bl64_dbg_lib_show_function "$@"
+  local kubeconfig="${1:-${BL64_LIB_DEFAULT}}"
+  local namespace="${2:-${BL64_LIB_DEFAULT}}"
+  local chart="${3:-${BL64_LIB_DEFAULT}}"
+  local source="${4:-${BL64_LIB_DEFAULT}}"
+
+  bl64_check_parameter 'kubeconfig' &&
+    bl64_check_parameter 'namespace' &&
+    bl64_check_parameter 'chart' &&
+    bl64_check_parameter 'source' &&
+    bl64_check_file "$kubeconfig" ||
+    return $?
+
+  shift
+  shift
+  shift
+  shift
+
+  bl64_hlm_run_helm \
+    upgrade \
+    "$chart" \
+    "$source" \
+    --kubeconfig="$kubeconfig" \
+    --namespace "$namespace" \
+    --timeout "$BL64_HLM_K8S_TIMEOUT" \
+    --atomic \
+    --cleanup-on-fail \
+    --install \
+    --wait \
+    "$@"
+}
+
+#######################################
+# Command wrapper with verbose, debug and common options
+#
+# * Trust no one. Ignore inherited config and use explicit
+#
+# Arguments:
+#   $@: arguments are passed as-is to the command
+# Outputs:
+#   STDOUT: command output
+#   STDERR: command stderr
+# Returns:
+#   command exit status
+#######################################
+function bl64_hlm_run_helm() {
+  bl64_dbg_lib_show_function "$@"
+  local verbosity=' '
+
+  bl64_check_parameters_none "$#" &&
+    bl64_check_module_setup "$BL64_HLM_MODULE" ||
+    return $?
+
+  bl64_dbg_lib_command_enabled && verbosity="$BL64_HLM_SET_DEBUG"
+
+  unset HELM_CACHE_HOME
+  unset HELM_CONFIG_HOME
+  unset HELM_DATA_HOME
+  unset HELM_DEBUG
+  unset HELM_DRIVER
+  unset HELM_DRIVER_SQL_CONNECTION_STRING
+  unset HELM_MAX_HISTORY
+  unset HELM_NAMESPACE
+  unset HELM_NO_PLUGINS
+  unset HELM_PLUGINS
+  unset HELM_REGISTRY_CONFIG
+  unset HELM_REPOSITORY_CACHE
+  unset HELM_REPOSITORY_CONFIG
+  unset HELM_KUBEAPISERVER
+  unset HELM_KUBECAFILE
+  unset HELM_KUBEASGROUPS
+  unset HELM_KUBEASUSER
+  unset HELM_KUBECONTEXT
+  unset HELM_KUBETOKEN
+
+  bl64_dbg_lib_trace_start
+  # shellcheck disable=SC2086
+  "$BL64_HLM_CMD_HELM" \
+    $verbosity \
+    "$@"
+  bl64_dbg_lib_trace_stop
+}