+++ src/bl64-k8s-env.bash
@@ -1,0 +1,21 @@
+#######################################
+# BashLib64 / Module / Globals / Interact with Kubernetes
+#
+# Version: 1.0.0
+#######################################
+
+# Optional module. Not enabled by default
+export BL64_K8S_MODULE="$BL64_LIB_VAR_OFF"
+
+export BL64_K8S_CMD_KUBECTL="$BL64_LIB_UNAVAILABLE"
+
+export BL64_K8S_SET_VERBOSE_NONE=''
+export BL64_K8S_SET_VERBOSE_NORMAL=''
+export BL64_K8S_SET_VERBOSE_DEBUG=''
+export BL64_K8S_SET_OUTPUT_JSON=''
+export BL64_K8S_SET_OUTPUT_YAML=''
+export BL64_K8S_SET_OUTPUT_TXT=''
+
+# Variables for external commands
+export POD_NAMESPACE
+export KUBECONFIG