+++ src/bl64-tm.bash
@@ -1,0 +1,39 @@
+#######################################
+# BashLib64 / Module / Functions / Manage date-time data
+#
+# Version: 1.0.0
+#######################################
+
+#######################################
+# Get current date-time in timestamp format
+#
+# * Format: DDMMYYHHMMSS
+#
+# Arguments:
+#   None
+# Outputs:
+#   STDOUT: formated string
+#   STDERR: command Error message
+# Returns:
+#   date exit status
+#######################################
+function bl64_tm_create_timestamp() {
+  "$BL64_OS_CMD_DATE" '+%d%m%Y%H%M%S'
+}
+
+#######################################
+# Get current date-time in file timestamp format
+#
+# * Format: DD:MM:YYYY-HH:MM:SS-TZ
+#
+# Arguments:
+#   None
+# Outputs:
+#   STDOUT: formated string
+#   STDERR: command Error message
+# Returns:
+#   date exit status
+#######################################
+function bl64_tm_create_timestamp_file() {
+  "$BL64_OS_CMD_DATE" '+%d:%m:%Y-%H:%M:%S-UTC%z'
+}