+++ src/bl64-txt-env.bash
@@ -1,14 +1,16 @@
 #######################################
 # BashLib64 / Module / Globals / Manipulate text files content
 #
-# Version: 1.1.0
+# Version: 1.2.0
 #######################################
 
 export BL64_TXT_MODULE="$BL64_LIB_VAR_OFF"
 
 export BL64_TXT_CMD_AWK=''
+export BL64_TXT_CMD_CUT=''
 export BL64_TXT_CMD_GAWK=''
 export BL64_TXT_CMD_GREP=''
 export BL64_TXT_CMD_SED=''
 export BL64_TXT_CMD_TR=''
 export BL64_TXT_CMD_BASE64=''
+export BL64_TXT_CMD_ENVSUBST=''